iManifestFileList component
Takes as input a manifest file, which itself contains a list of files to process, plus optionally a base path to prepend to each file path.
Two output connectors - one giving a flow of files which have been added or modified, the other a flow of files which have been removed.